# swagger/swagger.yaml
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "Tranquilify server API",
    "description" : "API for Tranquilify's server",
    "contact" : {
      "email" : "wagenheimj@moravian.edu"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "https://virtserver.swaggerhub.com/itchybarn/Headphones-Server/1.0.0",
    "description" : "SwaggerHub API Auto Mocking"
  } ],
  "tags" : [ {
    "name" : "auth",
    "description" : "Authentication"
  }, {
    "name" : "account",
    "description" : "Account management and creation"
  } ],
  "paths" : {
    "/login" : {
      "post" : {
        "tags" : [ "auth" ],
        "summary" : "user login",
        "description" : "Verifies a username and password. Returns a pre-auth token\nthat can be used to call /auth/send and /auth/verify\n",
        "operationId" : "loginUser",
        "requestBody" : {
          "description" : "Login credentials",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/LoginCredentials"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "credentials accepted",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TokenResponse"
                },
                "example" : {
                  "token" : "<pre-auth JWT>",
                  "token_type" : "Bearer",
                  "expires_in" : 300
                }
              }
            }
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "403" : {
            "description" : "invalid username or password",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                },
                "example" : {
                  "code" : "invalid_credentials",
                  "message" : "The username or password is incorrect"
                }
              }
            }
          },
          "429" : {
            "$ref" : "#/components/responses/RateLimited"
          }
        }
      }
    },
    "/user" : {
      "post" : {
        "tags" : [ "account" ],
        "summary" : "create new user",
        "description" : "Creates a new user with login credentials and an auth method",
        "operationId" : "createUser",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UserCreationComponent"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "user created"
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "409" : {
            "$ref" : "#/components/responses/Conflict"
          },
          "429" : {
            "$ref" : "#/components/responses/RateLimited"
          }
        }
      }
    },
    "/auth/send" : {
      "post" : {
        "tags" : [ "auth" ],
        "summary" : "send a code to the user",
        "description" : "Sends a one-time code to the auth method for the pre-auth user",
        "operationId" : "sendAuthCode",
        "responses" : {
          "200" : {
            "description" : "authentication sent successfully"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthorized"
          },
          "403" : {
            "description" : "unable to send the code to the auth method",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                },
                "example" : {
                  "code" : "code_delivery_failed",
                  "message" : "Unable to send a code to your auth method"
                }
              }
            }
          },
          "429" : {
            "$ref" : "#/components/responses/RateLimited"
          }
        },
        "security" : [ {
          "preAuthAuth" : [ ]
        } ]
      }
    },
    "/auth/verify" : {
      "post" : {
        "tags" : [ "auth" ],
        "summary" : "verify the user's code",
        "description" : "Verifies a one-time code. On success, provides the user with a full session token",
        "operationId" : "verifyAuthCode",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AuthCodeRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "authentication process complete",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TokenResponse"
                },
                "example" : {
                  "token" : "<access JWT>",
                  "token_type" : "Bearer",
                  "expires_in" : 3600
                }
              }
            }
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthorized"
          },
          "403" : {
            "$ref" : "#/components/responses/InvalidCode"
          },
          "429" : {
            "$ref" : "#/components/responses/RateLimited"
          }
        },
        "security" : [ {
          "preAuthAuth" : [ ]
        } ]
      }
    },
    "/logout" : {
      "post" : {
        "tags" : [ "auth" ],
        "summary" : "end the current session",
        "description" : "Revokes the caller's refresh token so the session cannot be renewed.\nThe access token is not invalidated and remains valid until its exp.\nIdempotent: unknown or already-revoked tokens still return 200.\n",
        "operationId" : "logout",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/LogoutRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "session invalidated"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthorized"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/refresh" : {
      "post" : {
        "tags" : [ "auth" ],
        "summary" : "renew tokens",
        "description" : "Exchanges a valid refresh token for a new access token AND a new refresh token.\nRotates the old refresh token: it is marked revoked and linked to its replacement.\nUnknown, revoked, or expired tokens return 401 with `invalid_refresh_token`.\nNo bearer auth required; the refresh token in the body is the credential.\n",
        "operationId" : "refreshTokens",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "new access and refresh tokens issued",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/LoginResponse"
                },
                "example" : {
                  "access_token" : "<access JWT>",
                  "refresh_token" : "<new opaque refresh token>",
                  "token_type" : "Bearer",
                  "expires_in" : 900
                }
              }
            }
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "description" : "refresh token is invalid, revoked, or expired",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                },
                "example" : {
                  "code" : "invalid_refresh_token",
                  "message" : "Refresh token is invalid, revoked, or expired."
                }
              }
            }
          },
          "429" : {
            "$ref" : "#/components/responses/RateLimited"
          }
        }
      }
    },
    "/username" : {
      "get" : {
        "tags" : [ "account" ],
        "summary" : "get current user's username",
        "description" : "Returns the username of the currently logged-in user.",
        "operationId" : "getCurrentUsername",
        "responses" : {
          "200" : {
            "description" : "authenticated user's username",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Username"
                }
              }
            }
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthorized"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/change-pass" : {
      "post" : {
        "tags" : [ "account" ],
        "summary" : "change password",
        "description" : "Changes password of logged-in user. Requires current password.\nMay require step up protection if auth_time is older than a certain window.\n",
        "operationId" : "changePassword",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "password changed"
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthorized"
          },
          "403" : {
            "description" : "current password incorrect, or step-up auth required",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                },
                "examples" : {
                  "wrongPassword" : {
                    "summary" : "the current password is incorrect",
                    "value" : {
                      "code" : "invalid_credentials",
                      "message" : "The current password is incorrect."
                    }
                  },
                  "stepUp" : {
                    "summary" : "session is too old; re-authentication required",
                    "value" : {
                      "code" : "step_up_required",
                      "message" : "Recent re-authentication is required to change your password."
                    }
                  }
                }
              }
            }
          },
          "429" : {
            "$ref" : "#/components/responses/RateLimited"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/reset-pass" : {
      "post" : {
        "tags" : [ "account" ],
        "summary" : "reset a forgotten password",
        "description" : "Resets the password for a logged-out user. The caller supplies the active\none-time code previously sent to their auth method via /auth/send, along with\nthe new password. If the code is valid, the password is updated.\nNo bearer auth is required; the one-time code is the credential.\nTo avoid revealing which usernames exist, an unknown username and an\nincorrect/expired code return the same error.\n",
        "operationId" : "resetPassword",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ResetPasswordRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "password reset"
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "403" : {
            "$ref" : "#/components/responses/InvalidCode"
          },
          "429" : {
            "$ref" : "#/components/responses/RateLimited"
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "UserCreationComponent" : {
        "required" : [ "auth_method", "login_credentials" ],
        "type" : "object",
        "properties" : {
          "login_credentials" : {
            "$ref" : "#/components/schemas/LoginCredentials"
          },
          "auth_method" : {
            "$ref" : "#/components/schemas/AuthMethod"
          }
        }
      },
      "LoginCredentials" : {
        "required" : [ "password", "username" ],
        "type" : "object",
        "properties" : {
          "username" : {
            "$ref" : "#/components/schemas/Username"
          },
          "password" : {
            "$ref" : "#/components/schemas/Password"
          }
        }
      },
      "ChangePasswordRequest" : {
        "required" : [ "current_password", "new_password" ],
        "type" : "object",
        "properties" : {
          "current_password" : {
            "$ref" : "#/components/schemas/Password"
          },
          "new_password" : {
            "$ref" : "#/components/schemas/Password"
          }
        }
      },
      "ResetPasswordRequest" : {
        "required" : [ "code", "new_password", "username" ],
        "type" : "object",
        "properties" : {
          "username" : {
            "$ref" : "#/components/schemas/Username"
          },
          "code" : {
            "$ref" : "#/components/schemas/AuthCode"
          },
          "new_password" : {
            "$ref" : "#/components/schemas/Password"
          }
        }
      },
      "AuthCodeRequest" : {
        "required" : [ "code" ],
        "type" : "object",
        "properties" : {
          "code" : {
            "$ref" : "#/components/schemas/AuthCode"
          }
        }
      },
      "LogoutRequest" : {
        "required" : [ "refresh_token" ],
        "type" : "object",
        "properties" : {
          "refresh_token" : {
            "type" : "string",
            "description" : "The raw refresh token belonging to the session to revoke"
          }
        }
      },
      "RefreshRequest" : {
        "required" : [ "refresh_token" ],
        "type" : "object",
        "properties" : {
          "refresh_token" : {
            "type" : "string",
            "description" : "The raw refresh token issued by /login or a previous /refresh"
          }
        }
      },
      "LoginResponse" : {
        "required" : [ "access_token", "expires_in", "refresh_token", "token_type" ],
        "type" : "object",
        "properties" : {
          "access_token" : {
            "type" : "string",
            "description" : "Short-lived JWT used as a bearer credential"
          },
          "refresh_token" : {
            "type" : "string",
            "description" : "Long-lived opaque token used with /refresh to obtain a new pair"
          },
          "token_type" : {
            "type" : "string",
            "example" : "Bearer"
          },
          "expires_in" : {
            "type" : "integer",
            "description" : "Seconds until the access_token expires",
            "example" : 900
          }
        },
        "description" : "Access + refresh token pair returned by /login and /refresh"
      },
      "TokenResponse" : {
        "required" : [ "expires_in", "token", "token_type" ],
        "type" : "object",
        "properties" : {
          "token" : {
            "type" : "string",
            "description" : "Signed JWT, schema description for per-scope info"
          },
          "token_type" : {
            "type" : "string",
            "example" : "Bearer"
          },
          "expires_in" : {
            "type" : "integer",
            "description" : "Seconds until expiration",
            "example" : 300
          }
        },
        "description" : "Shared token type for every bearer token. Only JWT's token_use differs"
      },
      "Username" : {
        "maxLength" : 32,
        "minLength" : 3,
        "pattern" : "^[a-zA-Z0-9_.-]+$",
        "type" : "string",
        "example" : "username"
      },
      "Password" : {
        "maxLength" : 72,
        "minLength" : 8,
        "type" : "string",
        "description" : "Stored only as a salted hash; never returned by the API.",
        "format" : "password"
      },
      "AuthMethod" : {
        "type" : "string",
        "description" : "Destination to which a one-time code will be sent",
        "example" : "1235557890"
      },
      "AuthCode" : {
        "pattern" : "^[0-9]{6}$",
        "type" : "string",
        "description" : "Six-digit one-time code",
        "example" : "123456"
      },
      "Error" : {
        "required" : [ "code", "message" ],
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "string",
            "description" : "Error code",
            "enum" : [ "validation_failed", "invalid_credentials", "invalid_token", "invalid_refresh_token", "code_incorrect", "code_expired", "code_delivery_failed", "step_up_required", "username_taken", "rate_limited" ]
          },
          "message" : {
            "type" : "string",
            "description" : "For display/logging only"
          },
          "errors" : {
            "type" : "array",
            "description" : "Per-field detail, present when code is validation_failed",
            "items" : {
              "required" : [ "field", "issue" ],
              "type" : "object",
              "properties" : {
                "field" : {
                  "type" : "string",
                  "example" : "new_password"
                },
                "issue" : {
                  "type" : "string",
                  "example" : "must be at least 8 characters"
                }
              }
            }
          }
        }
      }
    },
    "responses" : {
      "BadRequest" : {
        "description" : "bad input parameter",
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/Error"
            },
            "examples" : {
              "validation" : {
                "summary" : "one or more fields failed the validation",
                "value" : {
                  "code" : "validation_failed",
                  "message" : "Some fields are invalid",
                  "errors" : [ {
                    "field" : "current_password",
                    "issue" : "is required"
                  }, {
                    "field" : "new_password",
                    "issue" : "must be at least 8 characters"
                  } ]
                }
              }
            }
          }
        }
      },
      "Unauthorized" : {
        "description" : "incorrect token",
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/Error"
            },
            "examples" : {
              "invalidToken" : {
                "summary" : "issue with the token",
                "value" : {
                  "code" : "invalid_token",
                  "message" : "The token is missing or invalid"
                }
              }
            }
          }
        }
      },
      "RateLimited" : {
        "description" : "too many requests",
        "headers" : {
          "Retry-After" : {
            "description" : "wait before retrying",
            "style" : "simple",
            "explode" : false,
            "schema" : {
              "type" : "integer"
            }
          }
        },
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/Error"
            },
            "examples" : {
              "rateLimited" : {
                "summary" : "too many attempts",
                "value" : {
                  "code" : "rate_limited",
                  "message" : "Too many requests. Try again later."
                }
              }
            }
          }
        }
      },
      "InvalidCode" : {
        "description" : "the supplied one-time code is incorrect or has expired",
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/Error"
            },
            "examples" : {
              "invalidCode" : {
                "summary" : "wrong or expired one-time code",
                "value" : {
                  "code" : "code_incorrect",
                  "message" : "The code is incorrect or has expired."
                }
              }
            }
          }
        }
      },
      "Conflict" : {
        "description" : "the resource already exists",
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/Error"
            },
            "examples" : {
              "usernameTaken" : {
                "summary" : "username is already registered",
                "value" : {
                  "code" : "username_taken",
                  "message" : "That username is already taken"
                }
              }
            }
          }
        }
      }
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "description" : "Full session token",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      },
      "preAuthAuth" : {
        "type" : "http",
        "description" : "Pre-auth token from /login, used for /auth/send and /auth/verify",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}